Refactor user registration process to use registerAllTestUsers function and streamline global setup
