+++ client-web/package.json
@@ -15,24 +15,17 @@
     "prebuild": "rimraf dist",
     "build": "tsc",
     "circular-dependencies": "madge --circular dist",
-    "format": "prettier --write \"src/**/*.ts\" \"test/**/*.ts\"",
     "lint": "tsc --noEmit && eslint src/**/*.ts{,x}",
     "lint:prod": "tsc --noEmit && cross-env NODE_ENV=production eslint src/**/*.ts{,x}",
     "lint:fix": "tsc --noEmit && eslint src/**/*.ts{,x} --fix",
     "codegen": "graphql-codegen --config codegen.ts",
-    "test": "jest --config ./config/jest.config.js",
-    "test:ui": "jest --detectOpenHandles --runInBand --config ./test/config/jest.config.ts",
     "stop-process": "fuser 4000/tcp -k",
-    "test:watch": "jest --watch",
-    "test:cov": "jest --coverage",
-    "test:debug": "node --inspect-brk -r tsconfig-paths/register -r ts-node/register node_modules/.bin/jest --runInBand",
     "test:auth-playwright": "playwright test",
-    "register-users": "tsx ./register-users.ts"
+    "test:nightly": "playwright test --config=config/playwright.config.nightly.ts"
   },
   "dependencies": {
     "@alkemio/client-lib": "^0.32.0",
     "@alkemio/tests-lib": "workspace:*",
-    "@ory/kratos-client": "^1.3.8",
     "axios": "1.13.5",
     "graphql": "^16.6.0",
     "graphql-upload": "^17.0.0",
