+++ lib/src/index.ts
@@ -20,6 +20,7 @@ export * from "./scenario/registration/register-in-alkemio-or-fail";
 export * from "./scenario/registration/register-in-kratos-or-fail";
 export * from "./scenario/registration/send-kratos-flow";
 export * from "./scenario/registration/verify-in-kratos-or-fail";
+export * from "./scenario/registration/register-test-user";
 export * from "./config/test.configuration";
 export * from "./config/alkemio-test-config";
 export * from "./config/create-config-using-envvars";
