+++ server-api/package.json
@@ -64,7 +64,6 @@
   "dependencies": {
     "@alkemio/client-lib": "^0.36.0",
     "@alkemio/tests-lib": "workspace:*",
-    "@ory/kratos-client": "^1.3.8",
     "axios": "1.13.5",
     "graphql": "^16.10.0",
     "ws": "^8.18.0"
